feat(custom-sw): removed file which does not make sense
